coupon/requirements.txt
absl-py==1.4.0
aioredis==1.3.1
amqp==1.4.9
anyio==3.6.2
anyjson==0.3.3
APScheduler==3.9.1.post1
arabic-reshaper==3.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.5.2
asn1crypto==1.5.1
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
axios==0.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
billiard==3.3.0.23
bleach==6.0.0
Brotli==1.0.9
cachetools==5.2.1
cairocffi==1.5.1
catboost==1.2
celery==3.1.26.post2
certifi==2021.10.8
cffi==1.14.4
channels==3.0.4
channels-redis==3.3.1
chardet==5.1.0
charset-normalizer==3.0.1
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cmake==3.25.2
colorama==0.4.4
comm==0.1.3
constantly==15.1.0
contourpy==1.0.7
cramjam==2.6.2
cryptography==38.0.1
cssselect2==0.7.0
cycler==0.11.0
daphne==3.0.2
DateTimeRange==2.1.0
debugpy==1.6.7
decorator==4.4.2
decouple==0.0.7
deepface==0.0.79
defusedxml==0.7.1
Deprecated==1.2.13
distlib==0.3.6
Django==4.1.7
django-apscheduler==0.6.2
django-background-tasks==1.2.5
django-bower==5.2.0
django-celery==3.3.1
django-compat==1.0.15
django-cors-headers==4.1.0
django-crontab==0.7.1
django-dramatiq==0.11.2
django-genius @ file:///C:/Users/santhosh%20parthiban/Desktop/pip/dist/django-genius-1.0.0.tar.gz#sha256=96b818a20246d3e0f50354804c4bd4f0e33b30ea40afa4aadaf3b3bfb6d5f8c2
django-mailbox==4.8.2
django-organizations==2.1.0
django-react @ file:///C:/Users/santhosh%20parthiban/Desktop/pip/dist/django-react-1.0.0.tar.gz#sha256=b9303fe5344769c311a586114a5572896f5a286a1cf1271e790f89902b84b46e
django-script @ file:///C:/Users/santhosh%20parthiban/Desktop/pip/dist/django-script-1.0.1.tar.gz#sha256=bc79d27ad39a42601a0a753e65907bd601e5f1df5ba3e71226d248ec28120f37
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djongo==1.3.6
dlib==19.24.0
dnspython==2.2.1
docutils==0.20.1
dramatiq==1.13.0
environ==1.0
eventlet==0.33.2
executing==1.2.0
face-recognition==1.3.0
face-recognition-models==0.3.0
fastjsonschema==2.17.1
fastparquet==2023.4.0
filelock==3.12.1
fire==0.5.0
Flask==2.0.3
flatbuffers==23.1.21
fonttools==4.31.1
fqdn==1.5.1
fsspec==2023.5.0
future==0.18.2
fuzzywuzzy==0.18.0
gast==0.4.0
gdown==4.7.1
gmvault==1.9.1
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
google-pasta==0.2.0
googleapis-common-protos==1.58.0
graphviz==0.20.1
gravatar==0.1
greenlet==2.0.1
grpcio==1.54.0
gunicorn==20.1.0
h11==0.13.0
h2==4.1.0
h5py==3.8.0
hiredis==2.1.1
hpack==4.0.0
html5lib==1.1
httpcore==0.16.3
httplib2==0.21.0
httpx==0.23.3
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.1
imageio==2.29.0
imageio-ffmpeg==0.4.8
IMAPClient==0.13
importlib-metadata==6.6.0
importlib-resources==5.12.0
imutils==0.5.4
incremental==22.10.0
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.7
isoduration==20.11.0
itsdangerous==2.1.0
jaraco.classes==3.2.3
jax==0.4.10
jedi==0.18.2
Jinja2==3.0.3
joblib==1.2.0
jsonpointer==2.4
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.8
keras==2.12.0
keyring==24.2.0
kiwisolver==1.4.0
kombu==3.0.37
libclang==16.0.0
libgravatar==1.0.3
libravatar-py==1.0.0
lxml==4.9.2
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mbstrdecoder==1.1.2
mdurl==0.1.2
mediapipe==0.9.1.0
mistune==3.0.1
ml-dtypes==0.1.0
more-itertools==9.1.0
MouseInfo==0.1.3
moviepy==1.0.3
msgpack==1.0.4
mtcnn==0.1.1
mysql-connector-python==8.0.28
mysqlclient==2.1.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==2.7.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.3
oauthlib==3.2.2
opencv-contrib-python==4.7.0.72
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
opt-einsum==3.3.0
optional-django==0.3.0
oscrypto==1.3.0
outcome==1.1.0
overrides==7.3.1
packaging==21.3
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pdfkit==1.0.0
pickleshare==0.7.5
pika==1.3.1
Pillow==9.5.0
pkginfo==1.9.6
platformdirs==3.5.3
playsound==1.3.0
plotly==5.15.0
proglog==0.1.10
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.5
psycopg2==2.9.6
psycopg2-binary==2.9.6
pure-eval==0.2.2
py-gravatar==0.0.3
pyarrow==12.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.53
pycparser==2.20
pydyf==0.6.0
PyGetWindow==0.0.9
Pygments==2.15.1
PyHamcrest==2.0.2
pyHanko==0.17.2
pyhanko-certvalidator==0.20.1
PyJWT==2.6.0
pyLibravatar==1.7
pymongo==3.11.4
PyMsgBox==1.0.9
pyOpenSSL==22.0.0
pyparsing==3.0.7
pypdf==3.8.0
PyPDF2==1.26.0
pyperclip==1.8.2
pyphen==0.14.0
pypng==0.20220715.0
PyRect==0.1.4
pyrsistent==0.19.3
PyScreeze==0.1.28
PySocks==1.7.1
python-avatars==1.3.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==0.21.1
python-json-logger==2.0.7
pytweening==1.0.4
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pywhatkit==5.3
pywin32==306
pywin32-ctypes==0.2.2
pywinpty==2.0.10
PyYAML==6.0
pyzmail==1.0.3
pyzmq==25.1.0
qrcode==7.4.2
qtconsole==5.4.3
QtPy==2.3.1
react==4.3.0
readme-renderer==40.0
redis==4.3.4
reportlab==3.6.6
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rest-apscheduler==0.1.6
retina-face==0.0.13
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.4.2
rsa==4.9
scikit-learn==1.3.0
scipy==1.10.1
selenium==4.1.2
Send2Trash==1.8.2
service-identity==21.1.0
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.2.4
stack-data==0.6.2
svglib==1.5.1
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-hub==0.8.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.20.0
trio-websocket==0.9.2
twilio==7.16.0
twine==4.0.2
Twisted==22.8.0
twisted-iocpsupport==1.0.2
txaio==20.12.1
typepy==1.3.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.2
uri-template==1.3.0
uritemplate==4.1.1
uritools==4.0.1
urllib3==1.26.14
uuid==1.30
vine==5.0.0
virtualenv==20.23.0
virtualenvwrapper-win==1.2.7
wcwidth==0.2.5
weasyprint==58.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.0.3
whitenoise==6.4.0
widgetsnbextension==4.0.8
wikipedia==1.4.0
wincertstore==0.2
wkhtmltopdf==0.2
wrapt==1.14.1
wsproto==1.1.0
xhtml2pdf==0.2.10
zipp==3.15.0
zope.interface==5.2.0
zopfli==0.2.2
